Add support for ShellCheck rule exclusion config
This allows users to specify which rules to ignore for ShellCheck.
Shell check has an `-e` or `--exclude` option which takes a comma
separated list of rules to ignore when linting. Since Hound doesn't have
a strategy for user passed command flags, we create a Hound-only
configuration for ShellCheck.

The expected format is a Yaml file, like this:
```yaml
exclude:
  - SC2034
  - SC1068
```