Add SwiftLint
This requires a compiled binary (compiled on the same Docker image as
Heroku). There's some basic detection if the code is running on OSX vs
Linux and will call the appropriate binary.

In order to ugprade SwiftLint, we'll have to re-build it using Swift on
the architecture similar to Heroku's.